move InsertRewriteStatement's logic to SQLRewriteEngine
