+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/core/rewrite/SQLRewriteEngine.java
@@ -17,16 +17,22 @@
 
 package org.apache.shardingsphere.core.rewrite;
 
+import com.google.common.base.Optional;
+import com.google.common.base.Preconditions;
 import org.apache.shardingsphere.core.metadata.table.TableMetas;
+import org.apache.shardingsphere.core.optimize.segment.insert.InsertValueContext;
 import org.apache.shardingsphere.core.optimize.statement.SQLStatementContext;
+import org.apache.shardingsphere.core.optimize.statement.impl.InsertSQLStatementContext;
 import org.apache.shardingsphere.core.rewrite.builder.parameter.ParameterBuilder;
 import org.apache.shardingsphere.core.rewrite.builder.parameter.ParameterBuilderFactory;
 import org.apache.shardingsphere.core.rewrite.builder.sql.SQLBuilder;
 import org.apache.shardingsphere.core.rewrite.statement.RewriteStatement;
 import org.apache.shardingsphere.core.rewrite.statement.RewriteStatementFactory;
+import org.apache.shardingsphere.core.rewrite.statement.constant.EncryptDerivedColumnType;
+import org.apache.shardingsphere.core.rewrite.statement.constant.ShardingDerivedColumnType;
+import org.apache.shardingsphere.core.rewrite.token.SQLTokenGenerators;
 import org.apache.shardingsphere.core.rewrite.token.builder.BaseTokenGeneratorBuilder;
 import org.apache.shardingsphere.core.rewrite.token.builder.EncryptTokenGenerateBuilder;
-import org.apache.shardingsphere.core.rewrite.token.SQLTokenGenerators;
 import org.apache.shardingsphere.core.rewrite.token.builder.ShardingTokenGenerateBuilder;
 import org.apache.shardingsphere.core.rewrite.token.pojo.SQLToken;
 import org.apache.shardingsphere.core.route.SQLRouteResult;
@@ -39,8 +45,12 @@ import org.apache.shardingsphere.core.rule.BaseRule;
 import org.apache.shardingsphere.core.rule.EncryptRule;
 import org.apache.shardingsphere.core.rule.MasterSlaveRule;
 import org.apache.shardingsphere.core.rule.ShardingRule;
+import org.apache.shardingsphere.core.strategy.encrypt.EncryptTable;
+import org.apache.shardingsphere.spi.encrypt.ShardingEncryptor;
+import org.apache.shardingsphere.spi.encrypt.ShardingQueryAssistedEncryptor;
 
 import java.util.Collections;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
@@ -66,6 +76,10 @@ public final class SQLRewriteEngine {
                             final SQLRouteResult sqlRouteResult, final String sql, final List<Object> parameters, final boolean isQueryWithCipherColumn) {
         baseRule = shardingRule;
         rewriteStatement = RewriteStatementFactory.newInstance(shardingRule, sqlRouteResult);
+        if (sqlRouteResult.getSqlStatementContext() instanceof InsertSQLStatementContext) {
+            processGeneratedKey((InsertSQLStatementContext) sqlRouteResult.getSqlStatementContext(), sqlRouteResult.getGeneratedKey().orNull());
+            processEncrypt((InsertSQLStatementContext) sqlRouteResult.getSqlStatementContext(), shardingRule.getEncryptRule());
+        }
         parameterBuilder = ParameterBuilderFactory.newInstance(rewriteStatement, parameters, sqlRouteResult);
         sqlTokens = createSQLTokens(
                 tableMetas, sqlRouteResult.getShardingConditions(), sqlRouteResult.getGeneratedKey().orNull(), sqlRouteResult.getRoutingResult().isSingleRouting(), isQueryWithCipherColumn);
@@ -73,9 +87,12 @@ public final class SQLRewriteEngine {
     }
     
     public SQLRewriteEngine(final EncryptRule encryptRule, final TableMetas tableMetas,
-                            final SQLStatementContext encryptStatement, final String sql, final List<Object> parameters, final boolean isQueryWithCipherColumn) {
+                            final SQLStatementContext sqlStatementContext, final String sql, final List<Object> parameters, final boolean isQueryWithCipherColumn) {
         baseRule = encryptRule;
-        rewriteStatement = RewriteStatementFactory.newInstance(encryptRule, encryptStatement);
+        rewriteStatement = RewriteStatementFactory.newInstance(encryptRule, sqlStatementContext);
+        if (sqlStatementContext instanceof InsertSQLStatementContext) {
+            processEncrypt((InsertSQLStatementContext) sqlStatementContext, encryptRule);
+        }
         parameterBuilder = ParameterBuilderFactory.newInstance(rewriteStatement, parameters);
         sqlTokens = createSQLTokens(tableMetas, new ShardingConditions(Collections.<ShardingCondition>emptyList()), null, false, isQueryWithCipherColumn);
         sqlBuilder = new SQLBuilder(sql, sqlTokens);
@@ -87,6 +104,51 @@ public final class SQLRewriteEngine {
         parameterBuilder = ParameterBuilderFactory.newInstance(rewriteStatement, Collections.emptyList());
         sqlTokens = createSQLTokens(null, new ShardingConditions(Collections.<ShardingCondition>emptyList()), null, false, false);
         sqlBuilder = new SQLBuilder(sql, sqlTokens);
+    }
+    
+    private void processGeneratedKey(final InsertSQLStatementContext insertSQLStatementContext, final GeneratedKey generatedKey) {
+        if (null != generatedKey && generatedKey.isGenerated()) {
+            Iterator<Comparable<?>> generatedValues = generatedKey.getGeneratedValues().descendingIterator();
+            for (InsertValueContext each : insertSQLStatementContext.getInsertValueContexts()) {
+                each.appendValue(generatedValues.next(), ShardingDerivedColumnType.KEY_GEN);
+            }
+        }
+    }
+    
+    private void processEncrypt(final InsertSQLStatementContext insertSQLStatementContext, final EncryptRule encryptRule) {
+        String tableName = insertSQLStatementContext.getTablesContext().getSingleTableName();
+        Optional<EncryptTable> encryptTable = encryptRule.findEncryptTable(tableName);
+        if (!encryptTable.isPresent()) {
+            return;
+        }
+        for (String each : encryptTable.get().getLogicColumns()) {
+            Optional<ShardingEncryptor> shardingEncryptor = encryptRule.findShardingEncryptor(tableName, each);
+            if (shardingEncryptor.isPresent()) {
+                encryptInsertValues(insertSQLStatementContext, encryptRule, shardingEncryptor.get(), tableName, each);
+            }
+        }
+    }
+    
+    private void encryptInsertValues(final InsertSQLStatementContext insertSQLStatementContext, final EncryptRule encryptRule, final ShardingEncryptor shardingEncryptor,
+                                     final String tableName, final String encryptLogicColumnName) {
+        int columnIndex = insertSQLStatementContext.getColumnNames().indexOf(encryptLogicColumnName);
+        for (InsertValueContext each : insertSQLStatementContext.getInsertValueContexts()) {
+            encryptInsertValue(encryptRule, shardingEncryptor, tableName, columnIndex, each, encryptLogicColumnName);
+        }
+    }
+    
+    private void encryptInsertValue(final EncryptRule encryptRule, final ShardingEncryptor shardingEncryptor,
+                                    final String tableName, final int columnIndex, final InsertValueContext insertValueContext, final String encryptLogicColumnName) {
+        Object originalValue = insertValueContext.getValue(columnIndex);
+        insertValueContext.setValue(columnIndex, shardingEncryptor.encrypt(originalValue));
+        if (shardingEncryptor instanceof ShardingQueryAssistedEncryptor) {
+            Optional<String> assistedColumnName = encryptRule.findAssistedQueryColumn(tableName, encryptLogicColumnName);
+            Preconditions.checkArgument(assistedColumnName.isPresent(), "Can not find assisted query Column Name");
+            insertValueContext.appendValue(((ShardingQueryAssistedEncryptor) shardingEncryptor).queryAssistedEncrypt(originalValue.toString()), EncryptDerivedColumnType.ENCRYPT);
+        }
+        if (encryptRule.findPlainColumn(tableName, encryptLogicColumnName).isPresent()) {
+            insertValueContext.appendValue(originalValue, EncryptDerivedColumnType.ENCRYPT);
+        }
     }
     
     private List<SQLToken> createSQLTokens(final TableMetas tableMetas, final ShardingConditions shardingConditions,
